narrow down the resolver public interface
